requirements.txt
streamlit==1.24.0
tensorflow==2.12.0
numpy==1.23.5
opencv-python==4.7.0.72
pillow==9.5.0
scikit-learn==1.2.2
pandas==1.5.3
matplotlib==3.7.1
